apps/phoenix-cv/requirements.txt
# 🔥 Phoenix CV - Python Dependencies
# Production-ready requirements

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# AI & ML
google-generativeai==0.3.2

# Database
supabase==2.0.0

# Async & HTTP
aiofiles==23.2.1
httpx==0.25.2

# Configuration
python-dotenv==1.0.0

# Utilities
python-multipart==0.0.6
python-dateutil==2.8.2

# Document Processing  
python-docx==1.1.0
reportlab==4.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1